Moved UserException to correct folder
