+++ src/main/java/dev/accessaid/AccessAid/Ratings/service/RatingService.java
@@ -7,8 +7,8 @@ import dev.accessaid.AccessAid.Places.model.Place;
 import dev.accessaid.AccessAid.Ratings.exceptions.RatingNotFoundException;
 import dev.accessaid.AccessAid.Ratings.exceptions.RatingSaveException;
 import dev.accessaid.AccessAid.Ratings.model.Rating;
+import dev.accessaid.AccessAid.User.exceptions.UserNotFoundException;
 import dev.accessaid.AccessAid.User.model.User;
-import dev.accessaid.AccessAid.User.service.UserNotFoundException;
 
 public interface RatingService {
